improve : 마이페이지 디테일  (#51)
* [improve] 닉네임 오류 수정

* [fix] : Card 일부 잘림문제

* [improve]  lint type mismatch

+++ app/src/main/java/com/ssu/better/presentation/ui/challenge/approve/ChallengeApproveScreen.kt
@@ -224,7 +224,7 @@ fun ChallengeApproveContent(
                         modifier = Modifier.padding(start = 20.dp, top = 20.dp),
                         verticalAlignment = Alignment.CenterVertically,
                     ) {
-                        CircleRankProfile(score = userScore, modifier = Modifier.size(40.dp))
+                        CircleRankProfile(score = userScore.toLong(), modifier = Modifier.size(40.dp))
 
                         Spacer(modifier = Modifier.width(10.dp))
